Factor-out special cases for slice arrays.
The new implementation removes any special treatment of slices and greatly
simplifies the implementation.

All special 'views' for slice data have been removed but slice data is now kept
in per-slice lists. This means it is now more difficult to query a slice for a
particular subband's coefficients, but this was not considered an important
function before.