Add 'defaults' facility to serdes module (and remove defaults from fixeddicts)
This commit overhauls the way default values work when generating bitstreams.

Previously bitstream defaults were assigned by the vc2_fixeddicts themselves
(via the 'default' and 'default_factory' mechanism now removed in this commit).
This had the advantage that after creating a fixeddict you could see all of the
default values it contained. The disadvantage was that some fields could not be
given default values since the existance of that field depended on previous
bitstream structures. As a result, producing a minimal picture definition, for
example, required almost the entire stream be explicitly defined, making tests
excessively verbose.

This change introduces two key features to the serdes module:

* A lookup-table of default values may be passed into a serialiser. Since this
  table is separate from the provided bitstream specification it is not
  restricted to only containing values which exist in the bitstream being
  seriallised. This means that default values for all possible fields may be
  provided in advance and then used as required by a particular bitstream.
* The serialiser now automatically zero-pads bitarray and bytes entries. This
  behaviour means that padding data need no longer be provided with exactly the
  correct length in advance -- a recurring frustration in tests where the
  padding fields are not the focus of attention.

These changes mean that complete sequences and pictures may be constructed with
minimal boilerplate substantially easing the development of tests.