Make encoder obey level constraints.
Previously the encoder would only obey the sequence header constraints and
ignore all others. Now, the constraints on extended transform parameters are
also obeyed.

The reason this seemingly small change has produced such a large commit is that
this uncovered the sore issue of level constraints being free-form constraints.
In the general case this would require a sequence-wide constraint solver to
implement. Fortunately VC-2's level constraints are such that this is not
necessary however a test was required to prove that this is the case (and fail
should it ever cease to be so).

This in turn has led to more refinements to the way constraint solving for the
encoder is done more generally, with more attention being paid to levels
generally.