Tidy up pseudocode metadata module
Previously this module had the ability to label all kinds of spec-related
functions, data and types. Most of this functionality turned out not to be
useful and so this commit removes it. The version is now focused entirely on
logging pseudocode function derived functions.

The only minor improvement external external to the module is the new
make_pseudocode_traceback function which factors pseudocode traceback
construction logic out of two of the scripts into a single place.