Add checks and auto-fill for major_version/minor_version
The VC-2 spec specifies exactly how the major_version and minor_version fields
must be filled in based on which features are used (or not used) by a sequence.
For instance, if a stream uses certain preset frame rate indices it must have
a major_version of 3 whilst streams using plain-old low-delay coding should
have major_version set to 1.

Previously, the conformance software did not perform any of the above checks
when validating a bitstream. Furthermore, it also left choosing the
major_version and minor_version up to the user when generating decoder test
cases. This meant that it would be possible (in fact likely) that users would
generate non-conforming test cases!

This commit introduces checks in the conformance checking decoder to verify
that the correct major_version is set (and that minor_version is always 0).

This commit also removes the major_version/minor_version fields from the codec
features tables used by users to control the test case generator. Instead the
test case generator now automatically determines these values based on the
rules defined in the spec.

An (un)fortunate consequence of the above change is that a number of test cases
which previously incorrectly set major_version to 3 (or rather left it to
default to 3) began to fail. These tests were repaired to use the correct
number for the features used (resulting in a fair bit of noise in this diff,
sorry!).